# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add-event": {
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Some location"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2016-05-26"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": "12"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event  Added."
                                        },
                                        "Event": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Some location"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "Some Date"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": "12"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": "122"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "all events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "122"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some Name"
                                                    },
                                                    "location": {
                                                        "type": "string",
                                                        "example": "Some location"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "Some Date"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "12"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-11-08T06:56:27.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-11-08T06:56:27.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/event/{event}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "show event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID Book",
                        "example": 28
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "28"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some Name"
                                        },
                                        "location": {
                                            "type": "string",
                                            "example": "Some location"
                                        },
                                        "date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "Some Date"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "12"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-11-08T06:56:27.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-event/{event}": {
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "update event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID Book",
                        "example": 70
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some Name"
                                            },
                                            "location": {
                                                "type": "string",
                                                "example": "Ereven Armenia"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2016-05-26"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Updated."
                                        },
                                        "Event": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "70"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some Name"
                                                },
                                                "location": {
                                                    "type": "string",
                                                    "example": "Ereven Armenia"
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2016-05-26"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2023-11-08T06:56:27.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/delete-event/{event}": {
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "delete event",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "ID Book",
                        "example": 74
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event Deleted."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}